refactor and rename method
Refactor from_xrda and rename it from_array_with_metadata.  This is the
method with somewhat funny behaviour that was copied from
satpy.composites.ColormapCompositor.build_colormap.
It includes some behaviour that I cannot explain, such as cutting off
the last value of the colour palette when meanings are not provided.
I copied this behaviour as-is from Satpy.